Created initial Arduino sketch
Arduino sketch was created. This sketch will be responsible for
processing the information sent by the Python script and communicating
directly with the Arduino board and ultimately controling the RGB LED.

This initial sketch defined the redPin, bluePin, and greenPin of the LED
and set the pinMode of each to OUTPUT. These are pins 11, 10, and 9,
respectively.