Add serial port comm and defined officeStatus var.
Added serial port communication in void setup() to enable communication
with the Python script.

Added if conditional to check if information was recieved through the
serial port.

Defined the string variable officeStatus that will receive its value
from the Python script and will determine the color of the LED.

Set if statements to check the value of officeStatus to set the color
of the LED appropriately. The LED will turn red for officeStatus = "busy",
white for officeStatus = "not busy", and will turn off for "off".