Create userApp.py
userApp.py is the Python script used to build the desktop app used to
control the status of the LED.

A basic script was created to test the communication between this script
and the Arduino sketch and the Arduino Elego UNO R3 board.

This initial script asks the user to type B, N, or O to manipulate the
color of the LED (B indicates 'busy' and turns the LED red, N indicates
'not busy' and turns the LED white, O indicates 'off'.)

This was maninly a prove of consept and the script will continue to
evolve.